% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Prints result of the analysis data
For simple use always install "Doebioresearch" package using install.package('doebioresearch') after installing this package, Do analysis by just putting data in the recommemnded format.Code is as follos analysis(design, model, data, replication, factA, factB, factC), for design use ,  rcbd -->  for randomized complete block  design , rcbd2 --> for 2 factorial  rcbd , rcbd3 --> for 3 factorial rcbd, crd -- > for complete randmized design crd2 --> for 2 factorial crd, crd3 --> for 3 factorial crd, lsd --> for latin square design ,split -- > for split plot design, strip --> for strip plot design , for model  ,0 for no test ,1 for LSD test 2 for Duncan test,3 for HSD test for Data put Dataframe name ( as seen in global environment)for replication , Block for split plot, if block is not avalable put 0, for Factor A ,Put Treatment name incase of LSD , RCBD and CRD , factor B,Put Second factor Name otherwise put 0 if its absent .factor C,Put third factor Name otherwise put 0 if its absent,}
\usage{
analysis(design, model, data, rep, factA, factB, factC)
}
\description{
Prints result of the analysis data
For simple use always install "Doebioresearch" package using install.package('doebioresearch') after installing this package, Do analysis by just putting data in the recommemnded format.Code is as follos analysis(design, model, data, replication, factA, factB, factC), for design use ,  rcbd -->  for randomized complete block  design , rcbd2 --> for 2 factorial  rcbd , rcbd3 --> for 3 factorial rcbd, crd -- > for complete randmized design crd2 --> for 2 factorial crd, crd3 --> for 3 factorial crd, lsd --> for latin square design ,split -- > for split plot design, strip --> for strip plot design , for model  ,0 for no test ,1 for LSD test 2 for Duncan test,3 for HSD test for Data put Dataframe name ( as seen in global environment)for replication , Block for split plot, if block is not avalable put 0, for Factor A ,Put Treatment name incase of LSD , RCBD and CRD , factor B,Put Second factor Name otherwise put 0 if its absent .factor C,Put third factor Name otherwise put 0 if its absent,
}
