% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultotdoc.R
\name{resultdoc}
\alias{resultdoc}
\title{Title}
\usage{
resultdoc(filename, design, model, data, rep, factA, factB, factC)
}
\arguments{
\item{filename}{type as.doc}

\item{design}{design is blocking technique you used in the experiment}

\item{model}{0 for no test 1 for LSD 2 for Duncan 3 for HSD test}

\item{data}{data frame name}

\item{rep}{replication factor}

\item{factC}{}
}
\value{

}
\description{
Title
}
